Remove "prohibited restrictions" section.
Major change in this pull is to eliminate the section on prohibited restrictions, since it had no specific prohibitions any more. Added some of the language to the permitted restrictions, but dropped the references to the AC.

Removed two comments that are no longer relevant.

Also cleared up titles and added section #s. If we don't like the addition of section #s, we should refer to each rule as a "Rule" and fix the references to Sections 1-3 to Rules 1-*6*.