Suggestions, mainly to add clarity to open works
1.2 Acceptable conditions
This issue came up during the drafting process for the CERN OHL. We were careful that the licence didn't seek to grant rights (and hence try to impose conditions) on IPRs that didn't exist. For example, where, as in the UK, in many cases it's not a breach of intellectual property rights, and therefore does not necessitate a licence, to make a 3d item from a 2d design, it's not acceptable simply to say "this licence grants you the right to make a 3d item from a 2d design" because this presupposes that a licence is necessary. Instead, it's much better to say "this licence grants you the rights, where necessary, to make a 3d item from a 2d design". This is important for a number of reasons - A. It does not set expectations that the particular IPR exists. I have a particular worry that a variant of the 'where value then right' fallacy will be applied, and that a future court will apply the reasoning that this IPR must exist, otherwise the licensor would not have applied a licence to it. It also seeks to establish community norms. B. This is a subtle, but I believe important, point, so bear with me! A licence may say 'You may do whatever you like with this work, provided that if you distribute any part of it, the whole work containing that part must be distributed under the same licence'. This looks like a standard sharealike licence, but it may be the case that the 'any part' is so small, or is otherwise not covered by copyright (through fair use/fair dealing for example) that the distribution in this form would not require a licence. Clearly, therefore, the distribution of the subsequent work is fine and does not infringe copyright. However, in a non-compliant licence, the distributor may lose her right to the ORIGINAL work by distributing contrary to this licence. This might seem like a highly technical issue, but in fact is important in the interpretation of GPL (v2 in particular), and with the current state of Oracle v Google, this really needs considering and analysing carefully. I admit A is more of a drafting issue for individual contracts, but I believe B is important for the definition. 

1.2.1 I have conflicting views about this clause. On the one hand, it should not only respect the attribution of contributors and creators, but also of rights holders (this is more for legal reasons - if the creators assign their rights to the Free Software Foundation, for example, the FSF will find it easier to enforce its rights if it is acknowledged as the rightsholder in the attriubtion) and sponsors (Canonical, in the case of Ubuntu, for example). On the other hand, this clause can, as is recognised, get very onerous, and I am glad that wording is in there, although the problem is that the onerousness of the clause is not so much determined by what is in the licence, but by the work to which it is attached, and so sits uneasily in the definition. If you look at the attribution clause in Apache, for example, that's fine for software, but would be horribly onerous if you applied it to an open database. Same licence - great for software, not so great for databases. 

1.2.3 Amended for clarity. Great clause, BTW - deals with both 'or later' and licences like EUPL which permit relicensing. 

2.1.1 General, for discussion (I haven't amended this as I'm not sure that it should go here, or in the licence part, but it's important that the rightsholder does not derogate from his grant by licensing with one licence (e.g. copyright), and then refusing to grant a patent licence, or granting it on discriminatory terms. On the other hand the rightsholder should NOT be required to grant a licence to trademarks, provided that it's reasonably simple to use the work without infringing the trademark, or to remove references to it entirely in order to allow the full spectrum of use.

2.1.2. This may be sneaking a form or copyleft (which is supposed to be optional) in by the back door, I admit, but I feel strongly that where a database, in particular, is provided, and the provider has generated metadata, that the metadata should also be available. 

2.1.3 I wanted to clarify that not only must the data be available in bulk, but the unnecessary technological obstacles must also apply to the downloading (where appropriate). For example, the system is designed to throttle downloads of more than so many records to a lower speed.