Minor language proposal to improve conssitency.
There is no difference between the meaning of “shall” and “must” in
RFC2119 and there does not appear to be a reason for using them
interchangeably in the definition here, so I propose that a single term
is used instead of two.

Ken Adam’s “Manual of Style for Contract Drafting” recommends that
“shall” is used as a replacement for “has a duty to” while “must” is
used in all other situations. Therefore “must” is the more appropriate
term for this.