gramattical fix
the "and" and the comma in 1.2 were grammatically incorrect combinations. The second and (now the only one given my edit) was not followed by something that works as a complete sentence, and the different qualifiers make it not a consistent list that deserves comma separation. Although my result is a long sentence, I'm making this fix just to be okay for now. I would prefer to rewrite the sentence with the "should" part first and an "at least" added to the "must" part, but we've started voting on moving forward with 2.1 already.

I hope we can at least accept my grammatical fix.