cleaner approval process, overall revisions
It makes no sense for a "release candidate" phase to have all of: a delay before a vote, and a vote, and an announcement, and a wait period, and another vote. The summary of the text from my update is:

* consensus seems apparent (no issues outstanding)
* do immediate vote on release candidate (no delay, no extra announcements)
* announce the release candidate, have delay, accept typo fixes, *readily* move back to earlier stages to create newer release candidate if issues come up
* when through this without issues, then final vote

Basically, there's no reason to be so hesitant to get to release candidates, do them OFTEN, quickly, once known issues are done, and then don't push them through so hard.